Reorganized project_types into separate files
