+++ src/lib.rs
@@ -4,19 +4,18 @@ extern crate sfml;
 extern crate regex;
 extern crate openssl;
 
-
 pub mod utils;
 mod init;
 mod user;
 mod sequence;
 mod project_types;
-mod error;
+pub mod error;
 mod permissions;
 
 // Re-exports
 pub use init::*;
 pub use user::*;
 pub use sequence::*;
-pub use project_types::*;
-pub use error::*;
+//pub use project_types::*;
+//pub use error::*;
 pub use permissions::*;
