+++ src/project_types/user.rs
@@ -1,0 +1,8 @@
+
+
+#[derive(Clone, Debug, PartialEq, Eq, RustcEncodable, RustcDecodable)]
+pub struct User {
+    pub name: String,
+    pub public_key: String,
+}
+
